requirements.txt
Django==1.11.3
django-mptt==0.8.7
pytz==2017.2
